+++ packages/python/qsu/object/__init__.py
@@ -1,6 +1,7 @@
 from .objDeleteKeyByValue import objDeleteKeyByValue
 from .objFindItemRecursiveByKey import objFindItemRecursiveByKey
 from .objMergeNewKey import objMergeNewKey
+from .objPickBy import objPickBy
 from .objTo1d import objTo1d
 from .objToArray import objToArray
 from .objToPrettyStr import objToPrettyStr
@@ -11,6 +12,7 @@ __all__ = [
 	'objDeleteKeyByValue',
 	'objFindItemRecursiveByKey',
 	'objMergeNewKey',
+	'objPickBy',
 	'objTo1d',
 	'objToArray',
 	'objToPrettyStr',